#59: jersey-se integrated
