+++ poetryclub-core/src/main/java/com/github/skapral/poetryclub/core/scalar/ScalarConfigurable.java
@@ -26,7 +26,8 @@
 
 package com.github.skapral.poetryclub.core.scalar;
 
-import com.github.skapral.poetryclub.core.config.ConfigProperty;
+import com.github.skapral.config.ConfigProperty;
+
 
 /**
  * Scalar with configurable value.
