+++ poetryclub-service/src/main/java/com/github/skapral/poetryclub/service/config/Cp_GITHUB_API_KEY.java
@@ -1,6 +1,6 @@
 package com.github.skapral.poetryclub.service.config;
 
-import com.github.skapral.poetryclub.core.config.CpEnvironment;
+import com.github.skapral.config.CpEnvironment;
 
 /**
  * GITHUB_API_KEY property
